pyproject.toml
[tool.poetry]
name = "screw-player"
version = "0.0.1"
description = ""
authors = ["Jordan Wolfaardt"]

[tool.poetry.dependencies]
pydantic = "^1.10.4"
pydealer = "^1.4.0"
python = "^3.10"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
mypy = "^0.942"
pytest = "^7.0.0"
pyzmq = "^25.0.0"

[tool.black]
line-length = 99
target-version = ['py39']
